% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{calculatePeakCharacteristics,cpc_chrom-method}
\alias{calculatePeakCharacteristics,cpc_chrom-method}
\title{Internal method for calculation of the peak characteristics}
\usage{
\S4method{calculatePeakCharacteristics}{cpc_chrom}(x)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}
}
\value{
cpc_chrom object
}
\description{
This is an internal function for calculating the peak characteristics based 
on the processing results.
}
