% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_chrom.R
\name{chromFromTrace}
\alias{chromFromTrace}
\title{Create a cpc_chrom object from a chromatogram trace}
\usage{
chromFromTrace(st, trace, param = cpcChromParam())
}
\arguments{
\item{st}{Scantimes for the chromatogram. If not supplied will default to 1:length(trace).}

\item{trace}{Numeric vector containing the chromatogram trace to be processed.}

\item{param}{*cpcChromParam* object contaning the processing parameters.}
}
\value{

}
\description{
This function creates a *cpc_chrom* object from a chromatogram trace for 
external processing of chromatograms without XCMS.
}
