% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cpc-class}
\alias{cpc-class}
\title{cpc class}
\description{
cpc class
}
\section{Slots}{

\describe{
\item{\code{xd}}{Original \code{XCMSnExp} object.}

\item{\code{xdFilt}}{Filtered \code{XCMSnExp} object.}

\item{\code{pt}}{\code{data.frame} containing the original peak table from \code{xd}.}

\item{\code{cpt}}{\code{data.frame} containing the determined peak characteristics.}

\item{\code{fdef}}{Feature definitions (not used currently).}

\item{\code{fpeaks}}{Feature defining peaks (not used currently).}

\item{\code{procData}}{\code{list} containing processing data.}

\item{\code{procParams}}{\code{list} containing processing parameters.}
}}

