% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{hasPeakTable,cpc-method}
\alias{hasPeakTable,cpc-method}
\alias{hasCharacterizedPeakTable,cpc-method}
\alias{xdObj,cpc-method}
\alias{xdObj<-,cpc-method}
\alias{peaksToKeep,cpc-method}
\alias{getPeaklist,cpc-method}
\alias{setPeaklist<-,cpc-method}
\alias{cpt,cpc-method}
\alias{cpt<-,cpc-method}
\alias{getProcData,cpc-method}
\alias{setProcData<-,cpc-method}
\alias{addProcData<-,cpc-method}
\alias{getProcParams,cpc-method}
\alias{setProcParams<-,cpc-method}
\alias{addProcParams<-,cpc-method}
\alias{parsePeaklist,cpc-method}
\alias{parseFeatures,cpc-method}
\alias{filePaths,cpc-method}
\alias{processPeaks,cpc-method}
\alias{filterPeaks,cpc-method}
\alias{getChromatogram,cpc-method}
\alias{show,cpc-method}
\alias{summary,cpc-method}
\title{Method to check if a cpc object has a parsed peaktable from XCMS.}
\usage{
\S4method{hasPeakTable}{cpc}(x)

\S4method{hasCharacterizedPeakTable}{cpc}(x)

\S4method{xdObj}{cpc}(x)

\S4method{xdObj}{cpc}(x) <- value

\S4method{peaksToKeep}{cpc}(x, returnBoolean = FALSE)

\S4method{getPeaklist}{cpc}(x)

\S4method{setPeaklist}{cpc}(x) <- value

\S4method{cpt}{cpc}(x)

\S4method{cpt}{cpc}(x) <- value

\S4method{getProcData}{cpc}(x, value = NULL)

\S4method{setProcData}{cpc}(x) <- value

\S4method{addProcData}{cpc}(x) <- value

\S4method{getProcParams}{cpc}(x, value = NULL)

\S4method{setProcParams}{cpc}(x) <- value

\S4method{addProcParams}{cpc}(x) <- value

\S4method{parsePeaklist}{cpc}(x)

\S4method{parseFeatures}{cpc}(x)

\S4method{filePaths}{cpc}(x)

\S4method{processPeaks}{cpc}(x)

\S4method{filterPeaks}{cpc}(x)

\S4method{getChromatogram}{cpc}(x, id)

\S4method{show}{cpc}(x)

\S4method{summary}{cpc}(x)
}
\arguments{
\item{x}{A cpc object containing an \code{XCMSnExp} object}

\item{value}{Parameter name, if NULL it returns the entire slot (default = NULL)}

\item{returnBoolean}{}
}
\value{
Boolean

Boolean

\code{XCMSnExp} object

cpc object

Vector of peak indices that pass the filter criteria after peak characterization

A data.frame of the peaktable

A cpc object

A data.frame of the characterized peak table

A cpc object

Parameter values as a list()

cpc object

cpc object

Parameter values as a list()

cpc object

cpc object

cpc object

cpc object

cpc object

cpc object

cpc object

cpc_chrom object

cpc object

cpc object
}
\description{
This method will set the \code{sel_peaks} parameter in \code{procParams} to 
include only peaks that make up a part of a feature.
It will then set the pt slot to the subsetted peak table from the \code{XCMSnExp} object.

Prior to running this method the peak table must be parsed using the \code{parsePeaklist()} method.

This method is run after \code{processPeaks} in order remove the peaks that 
and creates a filtered XCMSnExp object in the xdFilt slot.

Prior to running this method the peak table must be parsed using the \code{parsePeaklist} method.
}
