\encoding{UTF-8}
\name{cpc-package}
\alias{cpc-package}
\alias{cpc}
\docType{package}
\title{
  Comprehensive Peak Characterization of LC/GC-MS Data
}
\description{
  Allows comprehensive peak characterization and filtering using 
  analytical chemistry parameters of peaks detected using XCMS.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Kristian Pirttilä, kristian.pirttila@ilk.uu.se

Maintainer: Kristian Pirttilä <kristian.pirttila@ilk.uu.se>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
