% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cpc_chrom-class}
\alias{cpc_chrom-class}
\title{Class used to process an XIC and calculate peak characteristics}
\description{
Class used to process the peaks found by XCMS. All smoothing and processing 
of a chromatogram is performed using this class.
}
\section{Slots}{

\describe{
\item{\code{id}}{\code{integer} value indicating which peak in the peak table is being processed.}

\item{\code{xic}}{Raw XIC trace}

\item{\code{d0}}{Smoothed XIC trace}

\item{\code{d2}}{Smoothed second derivative of the XIC trace}

\item{\code{st}}{Scantimes}

\item{\code{param}}{cpcParam object holding the processing parameters}

\item{\code{mzMeta}}{Metadata from the raw MS files}

\item{\code{procParams}}{\code{list} containing processing parameters. (deprecated)}

\item{\code{procData}}{\code{list} containing processing data.}

\item{\code{procResults}}{\code{list} containing the processing results for peak \code{id}.}

\item{\code{rawProcResults}}{\code{list} containing the raw processing results returned by \code{process_chromatogram}.}

\item{\code{results}}{\code{list} containing calculated peak characteristics for peak \code{id}.}
}}

