% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setProcParams<-,cpc_chrom-method}
\alias{setProcParams<-,cpc_chrom-method}
\alias{addProcParams<-,cpc_chrom-method}
\alias{getProcParams,cpc_chrom-method}
\alias{addProcData<-,cpc_chrom-method}
\alias{getProcData,cpc_chrom-method}
\alias{getMzRange,cpc_chrom-method}
\alias{getResults,cpc_chrom-method}
\alias{setResults<-,cpc_chrom-method}
\alias{setXIC<-,cpc_chrom-method}
\alias{plotPeak,cpc_chrom-method}
\alias{calculatePeakCharacteristics,cpc_chrom-method}
\alias{smoothChromatogram,cpc_chrom-method}
\alias{fitEMG,cpc_chrom-method}
\alias{processChromatogram,cpc_chrom-method}
\title{Setter method for the procParams slot in cpc_chrom-class}
\usage{
\S4method{setProcParams}{cpc_chrom}(x) <- value

\S4method{addProcParams}{cpc_chrom}(x) <- value

\S4method{getProcParams}{cpc_chrom}(x, value = NULL)

\S4method{addProcData}{cpc_chrom}(x) <- value

\S4method{getProcData}{cpc_chrom}(x, value = NULL)

\S4method{getMzRange}{cpc_chrom}(x)

\S4method{getResults}{cpc_chrom}(x)

\S4method{setResults}{cpc_chrom}(x) <- value

\S4method{setXIC}{cpc_chrom}(x) <- value

\S4method{plotPeak}{cpc_chrom}(x)

\S4method{calculatePeakCharacteristics}{cpc_chrom}(x)

\S4method{smoothChromatogram}{cpc_chrom}(x)

\S4method{fitEMG}{cpc_chrom}(x)

\S4method{processChromatogram}{cpc_chrom}(x)
}
\arguments{
\item{x}{}

\item{value}{}
}
\value{
cpc_chrom object with updated procParams

cpc_chrom object with updated procParams

Specified procParam value

cpc_chrom-class with updated procData

Specified procData value

Specified procParam value

Value of results slot in object (data.frame).

cpc_chrom object with updated results slot.

Specified procParam value

None.

cpc_chrom object

cpc_chrom object.

cpc_chrom object.

cpc_chrom object
}
\description{
Setter method for the procParams slot in cpc_chrom-class

Add method for the procParams slot in cpc_chrom-class (deprecated! setProcParams<- has the same functionality.)

Getter method for the procParams slot in cpc_chrom-class

Add method for the procData slot in cpc_chrom-class

Getter method for the procData slot in cpc_chrom-class

Getter method for the procParams$mz_range slot in cpc_chrom-class

Getter method for the results slot in cpc_chrom-class

Setter method for the results slot in cpc_chrom-class

Setter method for the XIC slot in cpc_chrom-class

Plot method that plots a panel with results of the peak characterization of the VIP.

Method for calculating the peak characteristics based on the results from the peak characterization.

Method for calculating smoothed XICs in a cpc_chrom object.

Method for fitting a series of EMG functions to a characterized XIC.

Main method for characterizing an XIC.
}
