% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cpc_raw-class}
\alias{cpc_raw-class}
\title{Class for handling the MS raw data backend}
\description{
Class to manage the MS raw data backend using the mzR R-package to parse the
raw data.
}
\section{Slots}{

\describe{
\item{\code{file_path}}{File path to the raw MS file}

\item{\code{header}}{Header of the raw MS file}

\item{\code{runInfo}}{Experiment information from the raw MS file}

\item{\code{intensity}}{\code{numeric} vector of the individual MS peak intensities}

\item{\code{mz}}{\code{numeric} vector of the individual MS peak m/z values}

\item{\code{scanidx}}{\code{integer} vector indicating last idx in \code{intensity} and \code{mz} for each scan}

\item{\code{scanrate}}{Scans per second}

\item{\code{scantime}}{Seconds per scan}

\item{\code{nscan}}{Number of scans}

\item{\code{backend}}{String defining the backend used (currently not used).}
}}

