% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getXIC,cpc_raw-method}
\alias{getXIC,cpc_raw-method}
\alias{parseMz,cpc_raw-method}
\alias{scantime,cpc_raw-method}
\title{Getter that extracts an XIC from the raw MS data.}
\usage{
\S4method{getXIC}{cpc_raw}(x, mzrange, scanrange = NULL, method = 1)

\S4method{parseMz}{cpc_raw}(x)

\S4method{scantime}{cpc_raw}(x)
}
\arguments{
\item{x}{A cpc_raw object}

\item{method}{}
}
\value{
XIC trace

cpc_raw object

Scantime vector
}
\description{
Getter that extracts an XIC from the raw MS data.

Parsing method for the MS raw data contained in the specified file

Getter method for the scantime vector in a cpc_raw object with parsed MS data
}
