% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cpcChromParam-class}
\alias{cpcChromParam-class}
\title{Class cpcChromParam}
\value{

}
\description{
The class *cpcChromParam* holds all necessary parameters to characterize the 
peaks in a chromatogram. Instances of this class should be created using the
constructor *cpcChromParam*. The class *cpcChromParam* extends the class 
*cpcProcParam*.
}
\section{Slots}{

\describe{
\item{\code{mz}}{Numeric holding the m/z value of the selected peak.}

\item{\code{p}}{Retention time of the selected peak.}

\item{\code{s}}{Peak standard deviation to be used for determining the smoothing windows size.}

\item{\code{mz_range}}{Numeric vector of length 2 holding the m/z range for extracting the ion trace.}

\item{\code{nscan}}{Integer holding the length in number of scans of the ion trace.}
}}

