% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setParam<-,cpcParam-method}
\alias{setParam<-,cpcParam-method}
\alias{getParam,cpcParam-method}
\title{Setter method for the parameter slot}
\usage{
\S4method{setParam}{cpcParam}(x) <- value

\S4method{getParam}{cpcParam}(x, param)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{value}{A named \code{list} with parameter values.}

\item{param}{A string defining which parameter should be returned.}
}
\value{
A \code{cpc_chrom} object

A parameter value.
}
\description{
Takes a named \code{list} with parameter values as argument. Each existing
parameter given in the list will be updated with the value in the list. The
parameter names should be the same as the slot names in the *cpcProcParam* 
and *cpcChromParam* classes.
}
