% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{cpt<-,cpc-method}
\alias{cpt<-,cpc-method}
\title{Setter method for the \code{cpt} slot in a \code{cpc} object}
\usage{
\S4method{cpt}{cpc}(x) <- value
}
\arguments{
\item{x}{A \code{cpc} object}

\item{value}{A \code{data.frame} of the characterized peak table}
}
\value{
A \code{cpc} object
}
\description{
Setter method for the \code{cpt} slot in a \code{cpc} object
}
