% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{Euclidean}
\alias{Euclidean}
\title{Euclidean Distance from Dimension Means}
\usage{
Euclidean(data)
}
\arguments{
\item{data}{an \code{n x 2} matrix or data frame.}
}
\value{
Returns a numeric vector of length 1.
}
\description{
Calculates two-dimensional Euclidean distance between all points and dimension means.
}
\examples{
data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)

Euclidean(data)

}
