% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_match}
\alias{fast_match}
\title{Matches a vector of integers in a second vector of integers}
\usage{
fast_match(v1, v2)
}
\description{
Matches a vector of integers in a second vector of integers
}
