% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{filePaths,cpc-method}
\alias{filePaths,cpc-method}
\title{Getter for the \code{procData$file_paths} slot in a \code{cpc} object}
\usage{
\S4method{filePaths}{cpc}(x)
}
\arguments{
\item{x}{A \code{cpc} object}
}
\value{
A \code{character} vector with file paths
}
\description{
Getter for the \code{procData$file_paths} slot in a \code{cpc} object
}
