% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{filteredObject,cpc-method}
\alias{filteredObject,cpc-method}
\title{Getter method that returns the filtered \code{XCMSnExp} object.}
\usage{
\S4method{filteredObject}{cpc}(x)
}
\arguments{
\item{x}{A \code{cpc} object}
}
\value{
A filtered \code{XCMSnExp} object
}
\description{
Getter method that returns the filtered \code{XCMSnExp} object.
}
