% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_height_bounds}
\alias{find_height_bounds}
\title{Title}
\usage{
find_height_bounds(
  y = NULL,
  apex = NULL,
  bl_bounds = NULL,
  peak_bounds = NULL,
  frac = NULL,
  exact = TRUE,
  id = NULL,
  debug = FALSE,
  plot = FALSE
)
}
\arguments{
\item{y}{}

\item{apex}{}

\item{bl_bounds}{}

\item{peak_bounds}{}

\item{frac}{}

\item{exact}{}

\item{id}{}

\item{debug}{}

\item{plot}{}
}
\value{

}
\description{
Title
}
