% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getMzRange,cpc_chrom-method}
\alias{getMzRange,cpc_chrom-method}
\title{Getter for mz range parameter}
\usage{
\S4method{getMzRange}{cpc_chrom}(x)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}
}
\value{
A \code{numeric} vector with mz range.
}
\description{
Getter for mz range parameter
}
