% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getParam,cpc-method}
\alias{getParam,cpc-method}
\title{Getter method for the process parameter object contained in the cpc
object.}
\usage{
\S4method{getParam}{cpc}(x, param)
}
\arguments{
\item{x}{A \code{cpc} object.}

\item{value}{A \code{character} vector indicating a parameter slot in the
\code{cpcParam} object.}
}
\value{
The parameter value.
}
\description{
Takes a \code{cpc} object and a single character vector indicating a slot in 
the \code{cpcParam} object.
}
