% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getProcData,cpc-method}
\alias{getProcData,cpc-method}
\title{Getter for the \code{procData} slot in a \code{cpc} object}
\usage{
\S4method{getProcData}{cpc}(x, value = NULL)
}
\arguments{
\item{x}{A \code{cpc} object}

\item{value}{Parameter name, if NULL it returns the entire slot (default = NULL)}
}
\value{
Named \code{list} holding the parameter values
}
\description{
Getter for the \code{procData} slot in a \code{cpc} object
}
