% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getProcData,cpc_chrom-method}
\alias{getProcData,cpc_chrom-method}
\title{Getter method for the procData slot}
\usage{
\S4method{getProcData}{cpc_chrom}(x, value = NULL)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{value}{A named \code{list} with data. If \code{NULL}, returns a named \code{list} with all process data.}
}
\value{
A named \code{list} with process data.
}
\description{
Getter method for the procData slot
}
