% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getProcParams,cpc_chrom-method}
\alias{getProcParams,cpc_chrom-method}
\title{Getter method for the procParams slot}
\usage{
\S4method{getProcParams}{cpc_chrom}(x, value = NULL)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{value}{A named \code{list} of parameters. If \code{NULL}, returns a named \code{list} of all process parameters.}
}
\value{
A named \code{list} of process parameters.
}
\description{
Getter method for the procParams slot
}
