% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getResults,cpc_chrom-method}
\alias{getResults,cpc_chrom-method}
\title{Getter method for the results slot}
\usage{
\S4method{getResults}{cpc_chrom}(x)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}
}
\value{
A \code{data.frame} with the results from the peak processing.
}
\description{
Getter method for the results slot
}
