% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getXIC,cpc_raw-method}
\alias{getXIC,cpc_raw-method}
\title{Method to generate an XIC from a \code{cpc_raw} object}
\usage{
\S4method{getXIC}{cpc_raw}(x, mzrange, scanrange = NULL, method = 1)
}
\arguments{
\item{x}{A \code{cpc_raw} object with parsed MS data}

\item{mzrange}{A \code{numeric} vector of length 2.}

\item{scanrange}{An \code{integer} vector of length 2.}

\item{method}{A single \code{integer} indicating which method to use.}
}
\value{
A \code{numeric} vector.
}
\description{
Method to generate an XIC from a \code{cpc_raw} object
}
