% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{hasCharacterizedPeakTable,cpc-method}
\alias{hasCharacterizedPeakTable,cpc-method}
\title{Method to check if a \code{cpc} object has a characterized peaktable}
\usage{
\S4method{hasCharacterizedPeakTable}{cpc}(x)
}
\arguments{
\item{x}{A \code{cpc} object}
}
\value{
A \code{logical} indicating if there is a cpc characterized peak table in the object
}
\description{
Method to check if a \code{cpc} object has a characterized peaktable
}
