% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{hasPeakTable,cpc-method}
\alias{hasPeakTable,cpc-method}
\title{Method to check if a \code{cpc} object has a parsed peaktable}
\usage{
\S4method{hasPeakTable}{cpc}(x)
}
\arguments{
\item{x}{A \code{cpc} object}
}
\value{
A \code{logical} indicating if there is a peak table in the object
}
\description{
Method to check if a \code{cpc} object has a parsed peaktable
}
