% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_2p_x}
\alias{interpolate_2p_x}
\title{Title}
\usage{
interpolate_2p_x(x = NULL, y = NULL, k = NULL, yval = NULL)
}
\arguments{
\item{x}{}

\item{y}{}

\item{k}{}

\item{yval}{}
}
\value{

}
\description{
Title
}
