% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_x}
\alias{interpolate_x}
\title{Intepolate x value between two points given a y value}
\usage{
interpolate_x(x = NULL, y = NULL, yval = NULL)
}
\arguments{
\item{x}{Vector of length 2 giving the x values}

\item{y}{Vector of length 2 giving the y values}

\item{yval}{y value}
}
\value{

}
\description{
Intepolate x value between two points given a y value
}
