% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_y}
\alias{interpolate_y}
\title{Intepolate y value between two points given an x value}
\usage{
interpolate_y(x = NULL, y = NULL, xval = NULL)
}
\arguments{
\item{x}{Vector of length 2 giving the x values}

\item{y}{Vector of length 2 giving the y values}

\item{xval}{x value}
}
\value{

}
\description{
Intepolate y value between two points given an x value
}
