% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{parseMz,cpc_raw-method}
\alias{parseMz,cpc_raw-method}
\title{Parsing method for the MS raw data contained in the specified file}
\usage{
\S4method{parseMz}{cpc_raw}(x)
}
\arguments{
\item{x}{A \code{cpc_raw} object}
}
\value{
A \code{cpc_raw} object
}
\description{
Parsing method for the MS raw data contained in the specified file
}
