% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{peak_integral}
\alias{peak_integral}
\title{Title}
\usage{
peak_integral(x = NULL, y = NULL, a = NULL, b = NULL)
}
\arguments{
\item{x}{}

\item{y}{}

\item{a}{}

\item{b}{}
}
\value{

}
\description{
Title
}
