% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_eval_eic.R
\name{plot_chromatograms}
\alias{plot_chromatograms}
\title{Plot chromatogram(s) with process results}
\usage{
plot_chromatograms(
  cpc,
  sel_peaks = NULL,
  out_path = NULL,
  out_prefix = NULL,
  dppm = 50,
  dmz = NULL,
  drt = NULL,
  device = "png",
  ...
)
}
\arguments{
\item{cpc}{A \code{cpc} object}

\item{sel_peaks}{An \code{integer} vector with peaks indices to be plotted}

\item{out_path}{Path to save plots. (Default: NULL)}

\item{out_prefix}{Prefix for filenames. (Default: NULL)}

\item{dppm}{ppm-value for generating XICs.}

\item{dmz}{m/z range for generating XICs (ignored in dppm is set)}

\item{drt}{Retention time range for generated XICs (if NULL entire chromatogram is presented)}

\item{device}{Graphics device to be used (Default: \code{png})}

\item{...}{Further arguments passed to graphics device}
}
\value{

}
\description{
This function is used to plot an evaluation panel with results from peak 
processing using the cpc-package.
}
