% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_eval_eic.R
\name{plot_eval_eic}
\alias{plot_eval_eic}
\title{Title}
\usage{
plot_eval_eic(
  xd = NULL,
  pks = NULL,
  cpt = NULL,
  out_path = NULL,
  prefix = NULL,
  ...
)
}
\arguments{
\item{xd}{}

\item{pks}{}

\item{cpt}{}

\item{out_path}{}

\item{prefix}{}

\item{...}{}
}
\value{

}
\description{
Title
}
