% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_eval_eic.R
\name{plotPeaks}
\alias{plotPeaks}
\title{Plotting of characterized peaks}
\usage{
plotPeaks(
  cpc,
  peakIdx,
  outPath = NULL,
  prefix = NULL,
  device = "png",
  plotEMG = FALSE,
  plotXCMS = TRUE,
  noPrompt = FALSE,
  annotation = character(1),
  ...
)
}
\arguments{
\item{cpc}{A \code{cpc} object}

\item{peakIdx}{\code{numeric} vector of peak indices to be plotted}

\item{outPath}{(optional) Path to save plots to}

\item{prefix}{(optional) Filename prefix, if NULL the filename of the raw data is used}

\item{device}{Graphics device to be used for saving plots (default: "png")}

\item{plotEMG}{\code{logical} indicating if the EMG fitting results should be plotted (default: FALSE)}

\item{plotXCMS}{\code{logical} indicating if the XCMS data should be included in the plot (default: TRUE)}

\item{noPrompt}{\code{logical} indicating if the prompt when many peaks are being plotted will be suppressed (default: FALSE)}

\item{annotation}{(Optional) A single string or a vector of strings for each plotted peak with additional annotations for the plot}

\item{...}{Additional parameters to be sent to the graphics device}
}
\value{
\code{NULL}
}
\description{
Allows visualization of the characterization results. Plots the chromatogram and the second derivative with peak boundaries illustrated.
}
