% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{scantime,cpc_raw-method}
\alias{scantime,cpc_raw-method}
\title{Getter method for the scantime vector in a \code{cpc_raw} object}
\usage{
\S4method{scantime}{cpc_raw}(x)
}
\arguments{
\item{x}{A \code{cpc_raw} object}
}
\value{
A \code{numeric} vector with scantimes
}
\description{
Getter method for the scantime vector in a \code{cpc_raw} object
}
