% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setParam<-,cpc-method}
\alias{setParam<-,cpc-method}
\title{Setter method for the process parameter object contained in the cpc
object.}
\usage{
\S4method{setParam}{cpc}(x) <- value
}
\arguments{
\item{x}{A \code{cpc} object.}

\item{value}{A named \code{list} of parameters.}
}
\value{
A \code{cpc} object.
}
\description{
Takes a named \code{list} of parameters or another \code{cpcParam} object as 
argument. The element names in the \code{list} must correspond to a slot name in the 
cpcProcParam object.
}
