% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setPeaklist<-,cpc-method}
\alias{setPeaklist<-,cpc-method}
\title{Setter method for the parsed peak table in a \code{cpc} object}
\usage{
\S4method{setPeaklist}{cpc}(x) <- value
}
\arguments{
\item{x}{A \code{cpc} object}

\item{value}{A \code{data.frame} with peak information}
}
\value{
A \code{cpc} object
}
\description{
Setter method for the parsed peak table in a \code{cpc} object
}
