% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setProcData<-,cpc-method}
\alias{setProcData<-,cpc-method}
\title{Setter for the \code{procData} slot in a \code{cpc} object

Takes a named \code{list} of parameters as argument. If a parameter name 
already exists, it is updated with the new value and if it does not already
exist, it is added.}
\usage{
\S4method{setProcData}{cpc}(x) <- value

\S4method{setProcData}{cpc}(x) <- value
}
\arguments{
\item{x}{A \code{cpc} object}

\item{value}{Named \code{list} of parameter values}
}
\value{
A \code{cpc} object

A \code{cpc} object
}
\description{
Setter for the \code{procData} slot in a \code{cpc} object

Takes a named \code{list} of parameters as argument. If a parameter name 
already exists, it is updated with the new value and if it does not already
exist, it is added.

Setter for the \code{procData} slot in a \code{cpc} object
}
