% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setProcData<-,cpc_chrom-method}
\alias{setProcData<-,cpc_chrom-method}
\title{Setter method for the procData slot}
\usage{
\S4method{setProcData}{cpc_chrom}(x) <- value
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{value}{A named \code{list} with data.}
}
\value{
A \code{cpc_chrom} object
}
\description{
Takes a named \code{list} with data as argument. If a data name 
already exists, it is updated with the new value and if it does not already
exist, it is added.
}
