% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setProcParams<-,cpc_chrom-method}
\alias{setProcParams<-,cpc_chrom-method}
\title{Setter method for the procParams slot}
\usage{
\S4method{setProcParams}{cpc_chrom}(x) <- value
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{value}{A named \code{list} of parameters.}
}
\value{
A \code{cpc_chrom} object
}
\description{
Takes a named \code{list} of parameters as argument. If a parameter name 
already exists, it is updated with the new value and if it does not already
exist, it is added.
}
