% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setResults<-,cpc_chrom-method}
\alias{setResults<-,cpc_chrom-method}
\title{Setter method for the results slot}
\usage{
\S4method{setResults}{cpc_chrom}(x) <- value
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}
}
\value{
A \code{cpc_chrom} object.
}
\description{
Setter method for the results slot
}
