% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{setXIC<-,cpc_chrom-method}
\alias{setXIC<-,cpc_chrom-method}
\title{Setter method for the XIC slot}
\usage{
\S4method{setXIC}{cpc_chrom}(x) <- value
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{value}{A \code{numeric} vector.}
}
\value{
A \code{cpc_chrom} object
}
\description{
Setter method for the XIC slot
}
