% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,cpcParam-method}
\alias{show,cpcParam-method}
\title{Show method for the cpcParam object}
\usage{
\S4method{show}{cpcParam}(object)
}
\arguments{
\item{object}{A \code{cpc} object}
}
\description{
Outputs a summary of the cpcParam object.
}
