% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{smoothChromatogram,cpc_chrom-method}
\alias{smoothChromatogram,cpc_chrom-method}
\title{Method for calculating smoothed XICs in a cpc_chrom object.}
\usage{
\S4method{smoothChromatogram}{cpc_chrom}(x)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}
}
\value{
A \code{cpc_chrom} object.
}
\description{
Method for calculating smoothed XICs in a cpc_chrom object.
}
