% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS.R
\name{SS}
\alias{SS}
\title{Sum-of-Squares Calculation}
\usage{
SS(data, ...)
}
\arguments{
\item{data}{a numeric vector or \code{n x k} matrix or data frame.}

\item{...}{arguments passed to \code{\link{dist}()}.}
}
\value{
Returns a numeric vector of length 1.
}
\description{
Calculates sums of squares for uni- or multi-dimensional numeric data using the
distance matrix.
}
\examples{
data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
SS(data)

}
