% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{xdObj,cpc-method}
\alias{xdObj,cpc-method}
\title{Getter method for the \code{XCMSnExp} object contained in a \code{cpc} object}
\usage{
\S4method{xdObj}{cpc}(x)
}
\arguments{
\item{x}{A \code{cpc} object}
}
\value{
A \code{XCMSnExp} object
}
\description{
Getter method for the \code{XCMSnExp} object contained in a \code{cpc} object
}
