% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{xdObj<-,cpc-method}
\alias{xdObj<-,cpc-method}
\title{Setter method for the \code{XCMSnExp} object slot in a \code{cpc} object}
\usage{
\S4method{xdObj}{cpc}(x) <- value
}
\arguments{
\item{x}{A \code{cpc} object}
}
\value{
A \code{cpc} object
}
\description{
Setter method for the \code{XCMSnExp} object slot in a \code{cpc} object
}
