% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_characterize_peaklist.R
\name{characterize_xcms_peaklist}
\alias{characterize_xcms_peaklist}
\title{Wrapper function for characterization of the peaks detected by XCMS}
\usage{
characterize_xcms_peaklist(xd, param = cpcProcParam())
}
\arguments{
\item{xd}{\code{XCMSnExp} object with peak information}

\item{param}{\code{cpcProcParam} object with process parameters}
}
\value{
\code{cpc} object
}
\description{
This wrapper function for the characterization of all peaks detected by
XCMS takes an *XCMSnExp* object and a *cpcProcParam* object as arguments
and will run over all peaks detected in all files or the subset defined in 
the *cpcProcParam* object. Before running this function, XCMS must be used
to detect chromatographic peaks in the data. Currently the available algorithm for peak
characterization is an adaption of the ApexTrack algorithm implemented in 
many instrumental software. In short, the algorithm proceeds by first 
detecting all peak apices present as negative minima craddled by inflection
points in the second derivative of the chromatogram. All detected peak apices
is then subjected to a baseline expansion algorithm in order to determine the
peak boundaries and baseline boundaries.

This function is called by \code{filter_xcms_peaklist} but can be called 
independently to only characterize the peaks in the \code{XCMSnExp}.
}
\seealso{
\link{filter_xcms_peaklist}, \link{cpc-class}, \link{cpcProcParam}
}
