% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_characterize_peaklist.R
\name{characterize_xcms_peaklist}
\alias{characterize_xcms_peaklist}
\title{Wrapper function for characterization of the peaks detected by XCMS}
\usage{
characterize_xcms_peaklist(xd, param = cpcProcParam())
}
\arguments{
\item{xd}{\code{XCMSnExp} object with peak information}

\item{param}{\code{cpcProcParam} object with process parameters}
}
\value{
\code{cpc} object
}
\description{
This function is called by \code{filter_xcms_peaklist} but can be called 
independently to only characterize the peaks in the \code{XCMSnExp}.

TODO: Add a description about the algorithms used.
}
\seealso{
\link{filter_xcms_peaklist}, \link{cpc-class}, \link{cpcProcParam}
}
