% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getPeaklist,cpc-method}
\alias{getPeaklist,cpc-method}
\alias{cpt,cpc-method}
\alias{getProcData,cpc-method}
\alias{setProcData<-,cpc-method}
\alias{getProcParams,cpc-method}
\alias{getParam,cpc-method}
\alias{setParam<-,cpc-method}
\alias{parsePeaklist,cpc-method}
\alias{getChromatogram,cpc-method}
\alias{processPeaks,cpc-method}
\alias{getFilterOutcomes,cpc-method}
\alias{filterPeaks,cpc-method}
\alias{getOriginalXCMS,cpc-method}
\alias{getFilteredXCMS,cpc-method}
\alias{getRemovedPeaks,cpc-method}
\alias{getRetainedPeaks,cpc-method}
\alias{show,cpc-method}
\title{Getter method for the parsed peak table in a \code{cpc} object}
\usage{
\S4method{getPeaklist}{cpc}(x)

\S4method{cpt}{cpc}(x)

\S4method{getProcData}{cpc}(x, value = NULL)

\S4method{setProcData}{cpc}(x) <- value

\S4method{setProcData}{cpc}(x) <- value

\S4method{getProcParams}{cpc}(x, value = NULL)

\S4method{getParam}{cpc}(x, param)

\S4method{setParam}{cpc}(x) <- value

\S4method{parsePeaklist}{cpc}(x)

\S4method{getChromatogram}{cpc}(x, id)

\S4method{processPeaks}{cpc}(x)

\S4method{getFilterOutcomes}{cpc}(x)

\S4method{filterPeaks}{cpc}(x)

\S4method{getOriginalXCMS}{cpc}(x)

\S4method{getFilteredXCMS}{cpc}(x)

\S4method{getRemovedPeaks}{cpc}(x)

\S4method{getRetainedPeaks}{cpc}(x)

\S4method{show}{cpc}(object)
}
\arguments{
\item{x}{A \code{cpc} object}

\item{value}{A named \code{list} of parameters.}

\item{id}{The peak ID}

\item{object}{A \code{cpc} object}
}
\value{
A \code{data.frame} of the peak table

A \code{data.frame} of the characterized peak table

Named \code{list} holding the parameter values

A \code{cpc} object

A \code{cpc} object

A named \code{list} of process parameters.

The parameter value.

A \code{cpc} object.

A \code{cpc} object

A \code{cpc_chrom} object

A \code{cpc} object

A \code{cpc} object

A \code{XCMSnExp} object

A \code{XCMSnExp} object

A vector of peak IDs

A vector of peak IDs
}
\description{
Method that returns the characterized peak list with all calculated peak characteristics.

Takes a \code{cpc} object and a single character vector indicating a slot in 
the \code{cpcParam} object.

Takes a named \code{list} of parameters or another \code{cpcParam} object as 
argument. The element names in the \code{list} must correspond to a slot name in the 
cpcProcParam object.

Parses the peak table contained in the \code{XCMSnExp} object and stores it 
in the \code{pt} slot.

Prior to running this method the peak table must be parsed using the 
\code{parsePeaklist} or \code{parseFeatures} methods.

This is the main method run on a \code{cpc} object to characterize the peaks detected by XCMS.
Prior to running this method the peak table must be parsed using the \code{parsePeaklist} method.

This method is run after \code{processPeaks} in order to remove the peaks 
that do not fulfill the criteria and creates a filtered \code{XCMSnExp} 
object and stores it in the \code{xdFilt} slot.

This method returns the unfiltered original \code{XCMSnExp} object.

This method returns the filtered XCMS object after running the CPC
algorithm.

This method will return a vector containing the peak IDs of the peaks that
were removed during the CPC filtering.

This method will return a vector containing the peak IDs of the peaks that
were *not* removed during the CPC filtering.

Outputs a summary of the cpc object.
}
