% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{hasPeakTable,cpc-method}
\alias{hasPeakTable,cpc-method}
\alias{hasCharacterizedPeakTable,cpc-method}
\alias{xdObj,cpc-method}
\alias{xdObj<-,cpc-method}
\alias{filteredObject,cpc-method}
\alias{peaksToKeep,cpc-method}
\alias{getPeaklist,cpc-method}
\alias{setPeaklist<-,cpc-method}
\alias{cpt,cpc-method}
\alias{cpt<-,cpc-method}
\alias{getProcData,cpc-method}
\alias{setProcData<-,cpc-method}
\alias{getProcParams,cpc-method}
\alias{setProcParams<-,cpc-method}
\alias{parsePeaklist,cpc-method}
\alias{parseFeatures,cpc-method}
\alias{filePaths,cpc-method}
\alias{getChromatogram,cpc-method}
\alias{processPeaks,cpc-method}
\alias{filterPeaks,cpc-method}
\alias{getOriginalXCMS,cpc-method}
\alias{getFilteredXCMS,cpc-method}
\alias{show,cpc-method}
\alias{summary,cpc-method}
\title{Method to check if a \code{cpc} object has a parsed peaktable}
\usage{
\S4method{hasPeakTable}{cpc}(x)

\S4method{hasCharacterizedPeakTable}{cpc}(x)

\S4method{xdObj}{cpc}(x)

\S4method{xdObj}{cpc}(x) <- value

\S4method{filteredObject}{cpc}(x)

\S4method{peaksToKeep}{cpc}(x, returnBoolean = FALSE)

\S4method{getPeaklist}{cpc}(x)

\S4method{setPeaklist}{cpc}(x) <- value

\S4method{cpt}{cpc}(x)

\S4method{cpt}{cpc}(x) <- value

\S4method{getProcData}{cpc}(x, value = NULL)

\S4method{setProcData}{cpc}(x) <- value

\S4method{setProcData}{cpc}(x) <- value

\S4method{getProcParams}{cpc}(x, value = NULL)

\S4method{setProcParams}{cpc}(x) <- value

\S4method{parsePeaklist}{cpc}(x)

\S4method{parseFeatures}{cpc}(x)

\S4method{filePaths}{cpc}(x)

\S4method{getChromatogram}{cpc}(x, id)

\S4method{processPeaks}{cpc}(x)

\S4method{filterPeaks}{cpc}(x)

\S4method{getOriginalXCMS}{cpc}(x)

\S4method{getFilteredXCMS}{cpc}(x)

\S4method{show}{cpc}(x)

\S4method{summary}{cpc}(x)
}
\arguments{
\item{x}{A \code{cpc} object}

\item{value}{A named \code{list} of parameters.}

\item{returnBoolean}{\code{boolean} indicating if ...}
}
\value{
A \link{boolean}

A \link{boolean}

A \code{XCMSnExp} object

A \code{cpc} object

A filtered \code{XCMSnExp} object

An \code{integer} vector of peak idx that pass the criteria

A \code{data.frame} of the peak table

A \code{cpc} object

A \code{data.frame} of the characterized peak table

A \code{cpc} object

Named \code{list} holding the parameter values

A \code{cpc} object

A \code{cpc} object

A named \code{list} of process parameters.

A \code{cpc} object.

A \code{cpc} object

A \code{cpc} object

A \code{character} vector with file paths

A \code{cpc_chrom} object

A \code{cpc} object

A \code{cpc} object
}
\description{
Takes a named \code{list} of parameters as argument. If a parameter name 
already exists, it is updated with the new value and if it does not already
exist, it is added.

Parses the peak table contained in the \code{XCMSnExp} object and stores it 
in the \code{pt} slot.

This method will set the \code{sel_peaks} parameter in \code{procParams} to 
include only peaks that make up a part of a feature. It will call 
\code{parsePeaklist} with the new peak subset.

Prior to running this method the peak table must be parsed using the 
\code{parsePeaklist} or \code{parseFeatures} methods.

This is the main method run on a \code{cpc} object to characterize the peaks detected by XCMS.
Prior to running this method the peak table must be parsed using the \code{parsePeaklist} method.

This method is run after \code{processPeaks} in order to remove the peaks 
that do not fulfill the criteria and creates a filtered \code{XCMSnExp} 
object and stores it in the \code{xdFilt} slot.
}
\seealso{
\link{parsePeaklist}

\link{parsePeaklist}, \link{parseFeatures}, \link{cpc_chrom-class}

\link{parsePeaklist}, \link{parseFeatures}

\link{processPeaks}
}
