\name{cpc-package}
\alias{cpc-package}
\alias{cpc}
\docType{package}
\title{Common principal component (CPC) analysis and applications}
\description{This package contains tools for common principal component (CPC) analysis and applications of the CPC model to various other statistical techniques. Functions are available for covariance matrix estimation, discriminant analysis and biplots, all under the CPC model.}
\details{
\tabular{ll}{
Package: \tab cpc\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2015-02-24\cr
License: \tab GPL-2\cr
}
}
\author{
Theo Pepler, Danie Uys, Sarel Steel

Maintainer: Theo Pepler <tpepler@sun.ac.za>
}
\references{
Pepler, P.T. (2014). The identification and application of common principal components. PhD dissertation in the Department of Statistics and Actuarial Science, Stellenbosch University.
}
\keyword{ package }
\seealso{
}
\examples{
}
