% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{plotPeak,cpc_chrom-method}
\alias{plotPeak,cpc_chrom-method}
\alias{processChromatogram,cpc_chrom-method}
\title{Plot method that plots a panel with results of the peak characterization of the VIP.}
\usage{
\S4method{plotPeak}{cpc_chrom}(x, plotEMG = TRUE, plotXCMS = TRUE, annotation = character(1))

\S4method{processChromatogram}{cpc_chrom}(x)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{plotEMG}{\code{logical} indicating if results from EMG deconvolution should be plotted (default: TRUE)}

\item{plotXCMS}{\code{logical} indicating if XCMS reported peak bounds should be plotted (default: TRUE)}

\item{annotation}{(Optional) A single string or a vector of strings for each plotted peak with additional annotations for the plot}
}
\value{
A \code{cpc_chrom} object.
}
\description{
This method will plot the processing result as a panel for review.

This method will perform the processing of a chromatogram on a \code{cpc_chrom}
object. This method is called repeatedly by \code{characterize_xcms_peaklist}.
}
