% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getProcParams,cpc_chrom-method}
\alias{getProcParams,cpc_chrom-method}
\alias{setProcParams<-,cpc_chrom-method}
\alias{getProcData,cpc_chrom-method}
\alias{setProcData<-,cpc_chrom-method}
\alias{getMzRange,cpc_chrom-method}
\alias{getResults,cpc_chrom-method}
\alias{setResults<-,cpc_chrom-method}
\alias{setXIC<-,cpc_chrom-method}
\alias{plotPeak,cpc_chrom-method}
\alias{calculatePeakCharacteristics,cpc_chrom-method}
\alias{smoothChromatogram,cpc_chrom-method}
\alias{fitEMG,cpc_chrom-method}
\alias{processChromatogram,cpc_chrom-method}
\title{Getter method for the procParams slot}
\usage{
\S4method{getProcParams}{cpc_chrom}(x, value = NULL)

\S4method{setProcParams}{cpc_chrom}(x) <- value

\S4method{getProcData}{cpc_chrom}(x, value = NULL)

\S4method{setProcData}{cpc_chrom}(x) <- value

\S4method{getMzRange}{cpc_chrom}(x)

\S4method{getResults}{cpc_chrom}(x)

\S4method{setResults}{cpc_chrom}(x) <- value

\S4method{setXIC}{cpc_chrom}(x) <- value

\S4method{plotPeak}{cpc_chrom}(x)

\S4method{calculatePeakCharacteristics}{cpc_chrom}(x)

\S4method{smoothChromatogram}{cpc_chrom}(x)

\S4method{fitEMG}{cpc_chrom}(x)

\S4method{processChromatogram}{cpc_chrom}(x)
}
\arguments{
\item{x}{A \code{cpc_chrom} object.}

\item{value}{A \code{numeric} vector.}
}
\value{
A named \code{list} of process parameters.

A \code{cpc_chrom} object

A named \code{list} with process data.

A \code{cpc_chrom} object

A \code{numeric} vector with mz range.

A \code{data.frame} with the results from the peak processing.

A \code{cpc_chrom} object.

A \code{cpc_chrom} object

cpc_chrom object

A \code{cpc_chrom} object.

A \code{cpc_chrom} object

A \code{cpc_chrom} object.
}
\description{
Takes a named \code{list} of parameters as argument. If a parameter name 
already exists, it is updated with the new value and if it does not already
exist, it is added.

Takes a named \code{list} with data as argument. If a data name 
already exists, it is updated with the new value and if it does not already
exist, it is added.

This method will plot the processing result as a panel for review.

This is an internal function for calculating the peak characteristics based 
on the processing results.

Internal method for fitting a series of EMG functions to the chromatogram for
deconvolution of the peak shapes.

This method will perform the processing of a chromatogram on a \code{cpc_chrom}
object. This method is called repeatedly by \code{characterize_xcms_peaklist}.
}
