% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{getXIC,cpc_raw-method}
\alias{getXIC,cpc_raw-method}
\alias{parseMz,cpc_raw-method}
\alias{scantime,cpc_raw-method}
\title{Method to generate an XIC from a \code{cpc_raw} object}
\usage{
\S4method{getXIC}{cpc_raw}(x, mzrange, scanrange = NULL, method = 1)

\S4method{parseMz}{cpc_raw}(x)

\S4method{scantime}{cpc_raw}(x)
}
\arguments{
\item{x}{A \code{cpc_raw} object}

\item{mzrange}{A \code{numeric} vector of length 2.}

\item{method}{A single \code{integer} indicating which method to use.}

\item{scanrage}{An \code{integer} vector of length 2.}
}
\value{
A \code{numeric} vector.

A \code{cpc_raw} object

A \code{numeric} vector with scantimes
}
\description{
Method to generate an XIC from a \code{cpc_raw} object

Parsing method for the MS raw data contained in the specified file

Getter method for the scantime vector in a cpc_raw object with parsed MS data
}
