% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cpcChromParam-class}
\alias{cpcChromParam-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{mz}}{numericOrNULL.}

\item{\code{p}}{numericOrNULL.}

\item{\code{s}}{numericOrNULL.}

\item{\code{mz_range}}{numericOrNULL.}

\item{\code{nscan}}{numericOrNULL.}
}}

