% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cpcProcParam-class}
\alias{cpcProcParam-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{ppm}}{numericOrNULL.}

\item{\code{min_pts}}{numericOrNULL.}

\item{\code{min_inf_width}}{numericOrNULL.}

\item{\code{min_sn}}{numericOrNULL.}

\item{\code{min_frac}}{numericOrNULL.}

\item{\code{min_intensity}}{numericOrNULL.}

\item{\code{min_shoulder_pts}}{numericOrNULL.}

\item{\code{min_rounded_pts}}{numericOrNULL.}

\item{\code{interval_tf}}{numericOrNULL.}

\item{\code{min_fwhm}}{numericOrNULL.}

\item{\code{min_w}}{numericOrNULL.}

\item{\code{max_w}}{numericOrNULL.}

\item{\code{smooth_method}}{character.}

\item{\code{smooth_times}}{numericOrNULL.}

\item{\code{smooth_win}}{numericOrNULL.}

\item{\code{max_sigma}}{numericOrNULL.}

\item{\code{fit_emg}}{logical.}

\item{\code{sel_peaks}}{numericOrNULL.}

\item{\code{sel_files}}{numericOrNULL.}

\item{\code{verbose_output}}{logical.}

\item{\code{save_all}}{logical.}

\item{\code{plot}}{logical.}
}}

