% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_param.R
\name{cpcProcParam}
\alias{cpcProcParam}
\title{Constructor for a cpcProcParam object}
\usage{
cpcProcParam(
  ppm = 50,
  min_pts = 7L,
  min_inf_width = 3,
  min_sn = 10,
  min_frac = 0.5,
  min_intensity = 1000L,
  smooth_method = "savgol",
  smooth_times = 2L,
  smooth_win = NULL,
  max_sigma = NULL,
  fit_emg = FALSE,
  sel_peaks = NULL,
  sel_files = NULL,
  verbose_output = FALSE,
  plot = FALSE
)
}
\arguments{
\item{ppm}{The \code{ppm} range for generating XICs}

\item{min_pts}{Filter criteria for minimum points between peak boundaries}

\item{min_inf_width}{Filter criteria for minimum points between peak inflection points. Should be set >=3 to avoid issues.}

\item{min_sn}{Filter criteria for minimum signal-to-noise ratio.}

\item{min_frac}{Filter criteria for minimum samples a peak is found in. This is currently not used as it is only applicable to feature filtering.}

\item{min_intensity}{Filter criteria for minimum peak area.}

\item{smooth_method}{Smoothing method used during processing. Available: "savgol" for Savitzky-Golay smoothing or "mean" for moving mean smoothing.}

\item{smooth_times}{Number of smooth iterations,}

\item{smooth_win}{Width of the smoothing function. If set to NULL it will be determined from the peak sigma values in the XCMSnExp object.}

\item{max_sigma}{Largest allowable sigmal value used to determine the peak widths from the XCMSnExp object.}

\item{fit_emg}{\code{logical} indicating if EMG deconvolution should be performed. Note that this adds a significant amount of time to the processing but gives better estimates of peak characteristics.}

\item{sel_peaks}{\code{integer} vector indicating a subset of peak indices to be processed from the XCMSnExp object. If NULL all peaks are processed.}

\item{sel_files}{\code{integer} vector indicating a subset of file indices to be processed from the XCMSnExp object. If NULL all files are processed.}

\item{verbose_output}{\code{logical} value indicating if output is to be given during processing.}

\item{plot}{\code{logical} indicating if the results should be plotted for each peak}
}
\value{
A \code{cpcProcParam} object
}
\description{
These objects are used to set the parameters for the peak processing. Filter 
criteria are used to filter out peaks that do not meet the criteria.
}
