% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_filter_xcms_peaklist.R
\name{filter_xcms_peaklist}
\alias{filter_xcms_peaklist}
\title{Wrapper function for filter the peaks detected by XCMS}
\usage{
filter_xcms_peaklist(
  xd,
  return_type = c("xcms", "cpc"),
  param = cpcProcParam()
)
}
\arguments{
\item{xd}{\code{XCMSnExp} object}

\item{return_type}{String indicating if the filtered XCMS object ('xcms') or the CPC object ('cpc') should be returned.}

\item{param}{\code{cpcProcParam} object with process parameters}
}
\value{
Either an \code{XCMSnExp} object or a \code{cpc} object.
}
\description{
Use this function to filter the peaks detected by XCMS based on width, size, 
and signal-to-noise. The supplied \code{XCMSnExp} object should have peak
information contained in it. Notice that any feature definitions and retention
time alignment information is removed in the process so these processing
steps will have to be performed again after running this function.
}
\seealso{
\link{characterize_xcms_peaklist}
}
